feat(humidity): add mqtt connector to save samples
