+++ main.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/homenitor/back/adapters"
 	"github.com/homenitor/back/app"
-	"github.com/homenitor/back/ports"
+	"github.com/homenitor/back/ports/mqtt"
 	"github.com/homenitor/back/ports/web"
 )
 
@@ -23,11 +23,13 @@ func main() {
 		mqttHost = "127.0.0.1"
 	}
 
-	mqttServer, err := ports.NewMQTTServer(mqttHost, 1883, service, logging)
+	mqttServer, err := mqtt.NewServer(mqttHost, 1883, service, logging)
 	if err != nil {
 		panic(err)
 	}
 
 	mqttServer.SubscribeToRoomTemperature("livingroom")
+	mqttServer.SubscribeToRoomHumidity("livingroom")
+
 	web.Start(service, logging)
 }
